Reorganize immersed boundary implementation + Field for bottom_height
